+++ Cadmus.Import.Proteus/CadmusEntrySetContextPatcher.cs
@@ -1,5 +1,6 @@
 ﻿using Cadmus.Core;
 using Fusi.Tools.Configuration;
+using Proteus.Core;
 using Proteus.Core.Regions;
 using Proteus.Entries.Export;
 using System;
@@ -15,9 +16,10 @@ namespace Cadmus.Import.Proteus;
 /// </summary>
 [Tag("it.vedph.entry-set-context-patcher.cadmus")]
 public class CadmusEntrySetContextPatcher : EntrySetContextPatcher,
-    IEntrySetContextPatcher
+    IEntrySetContextPatcher, IConfigurable<CadmusEntrySetContextPatcherOptions>
 {
     private readonly StandardItemSortKeyBuilder _sortKeyBuilder;
+    private CadmusEntrySetContextPatcherOptions _options = new();
 
     /// <summary>
     /// Initializes a new instance of the <see cref="CadmusEntrySetContextPatcher"/>
@@ -26,6 +28,18 @@ public class CadmusEntrySetContextPatcher : EntrySetContextPatcher,
     public CadmusEntrySetContextPatcher()
     {
         _sortKeyBuilder = new StandardItemSortKeyBuilder();
+    }
+
+    /// <summary>
+    /// Configures this patcher with the specified options.
+    /// </summary>
+    /// <param name="options">The options.</param>
+    /// <exception cref="ArgumentNullException">options</exception>
+    public void Configure(CadmusEntrySetContextPatcherOptions options)
+    {
+        ArgumentNullException.ThrowIfNull(options);
+        ConfigureBaseOptions(options);
+        _options = options;
     }
 
     /// <summary>
@@ -38,6 +52,14 @@ public class CadmusEntrySetContextPatcher : EntrySetContextPatcher,
         ArgumentNullException.ThrowIfNull(context);
 
         CadmusEntrySetContext ctx = (CadmusEntrySetContext)context;
+
+        // remove empty items if not allowed
+        if (!_options.AllowEmptyItems)
+        {
+            ctx.Items.RemoveAll(item => item.Parts.Count == 0);
+        }
+
+        // patch items
         foreach (IItem item in ctx.Items)
         {
             if (string.IsNullOrEmpty(item.FacetId)) item.FacetId = "default";
@@ -57,4 +79,21 @@ public class CadmusEntrySetContextPatcher : EntrySetContextPatcher,
         }
         return Task.CompletedTask;
     }
+}
+
+/// <summary>
+/// Options for <see cref="CadmusEntrySetContextPatcher"/>.
+/// </summary>
+public class CadmusEntrySetContextPatcherOptions : DisabledOptions
+{
+    /// <summary>
+    /// True to allow empty items in the context; otherwise, false. Default is
+    /// false. When empty items are not allowed, any item without parts is
+    /// removed by the patcher.
+    /// </summary>
+    /// <remarks>Typically when importing items from CSV an item is created
+    /// for each row; but this happens also for the first empty row, which
+    /// signals the end of data. In this case this would create a last empty
+    /// item, which usually is unwanted.</remarks>
+    public bool AllowEmptyItems { get; set; }
 }